# NO-ISSUE: Speed up test database creation