@jhernand 2022-11-14
This is not really relevant, we are already using `tmpfs` from the Makefile, which provides similar behaviour.